move search api to correct package for code generation
