+++ boudicca.base/search-api/src/main/kotlin/base/boudicca/api/search/model/Filters.kt
@@ -1,4 +1,4 @@
-package base.boudicca.search.model
+package base.boudicca.api.search.model
 
 @Deprecated("use FiltersQueryDTO/FiltersResultDTO endpoint")
 data class Filters(
