+++ boudicca.base/search-api/src/main/kotlin/base/boudicca/api/search/model/SearchResultDTO.kt
@@ -1,4 +1,4 @@
-package base.boudicca.search.model
+package base.boudicca.api.search.model
 
 import base.boudicca.model.Event
