+++ boudicca.base/search/src/main/kotlin/base/boudicca/search/service/query/evaluator/NoopEvaluator.kt
@@ -1,6 +1,6 @@
 package base.boudicca.search.service.query.evaluator
 
-import base.boudicca.search.model.ResultDTO
+import base.boudicca.api.search.model.ResultDTO
 import base.boudicca.search.service.query.Evaluator
 import base.boudicca.search.service.query.Expression
 import base.boudicca.search.service.query.Page
