+++ boudicca.base/search/src/main/kotlin/base/boudicca/search/service/query/evaluator/SimpleEvaluator.kt
@@ -1,7 +1,7 @@
 package base.boudicca.search.service.query.evaluator
 
 import base.boudicca.model.Entry
-import base.boudicca.search.model.ResultDTO
+import base.boudicca.api.search.model.ResultDTO
 import base.boudicca.search.service.query.*
 import base.boudicca.search.service.util.Utils
 import java.time.LocalDate
